backend/requirements.txt
fastapi
uvicorn
pandas
numpy
joblib
xgboost
pydantic
httpx
scikit-learn
python-dotenv
google-generativeai
